split utils.py into subfiles, fix test_pre.py
