+++ rom_operator_inference/pre/_finite_difference.py
