+++ rom_operator_inference/utils/__init__.py
@@ -1,0 +1,5 @@
+# utils/__init__.py
+"""Utility functions for Operator Inference."""
+
+from ._solver import *
+from ._kronecker import *
